feat(upload): very ugly abstracted upload
