+++ components/sidebarlayout.tsx
@@ -40,6 +40,7 @@ import {
   parseCooking,
   parseFishing,
   parseCrafting,
+  parseMuseum,
 } from "../utils";
 
 import Notification from "./notification";
@@ -50,6 +51,7 @@ import * as Popover from "@radix-ui/react-popover";
 import LoginModal from "./modals/login";
 import Example from "./popup";
 import CreditsModal from "./modals/credits";
+import { parseSaveFile } from "../utils/file";
 
 const semVerGte = require("semver/functions/gte");
 
@@ -125,145 +127,18 @@ const SidebarLayout = ({
     // ex: reader.onloadstart, reader.onprogress, and finally reader.onload when its finished.
 
     reader.onload = async function (event) {
-      // console.log(event.target?.result);
-      console.log("Parsing XML...");
-      const start = performance.now();
-      const parser = new XMLParser({ ignoreAttributes: false });
-      const jsonObj = parser.parse(event.target?.result as string);
-      console.log(
-        "Parsed XML in",
-        (performance.now() - start).toFixed(3),
-        "ms"
-      );
-
-      // check the version number of the SDV save file
       try {
-        const gameVersion: string = jsonObj.SaveGame.gameVersion;
-
-        // make sure the game version is at least 1.5.0
-        // we can modify this later if we decide to support older versions of SDV
-        if (!semVerGte(gameVersion, "1.5.0")) {
-          setErrorMSG(
-            "Please upload a save file from version 1.5.0 or newer. If you would like to request support for an older version, please contact us on Discord or open an issue on Github."
-          );
-          setShowErrorNotification(true);
-          console.log("Exiting...");
-          return;
+        const { success, timeTaken } = await parseSaveFile(
+          event.target?.result
+        );
+        if (success) {
+          setShowNotification(true);
+          setCompletedTime(timeTaken);
         }
       } catch (e) {
-        if (e instanceof TypeError) {
-          setErrorMSG(
-            "Invalid File Uploaded. Couldn't find Game Version, please upload a Stardew Valley save file."
-          );
-          setShowErrorNotification(true);
-          console.log("Exiting...");
-          return;
-        }
+        setErrorMSG(e as string);
+        setShowErrorNotification(true);
       }
-      console.log("Parsing information...");
-      // General Information
-      const { name, timePlayed, farmInfo } = parseGeneral(jsonObj);
-      const moneyEarned = parseMoney(jsonObj);
-      const { levels, maxLevelCount } = parseSkills(jsonObj);
-      const questsCompleted = parseQuests(jsonObj);
-      const { stardrops, stardropsCount } = parseStardrops(jsonObj);
-
-      // Fishing
-      const { allFish, totalCaught, uniqueCaught } = parseFishing(jsonObj);
-
-      // Cooking
-      const { cookedCount, knownCount, allRecipes } = parseCooking(jsonObj);
-      // Crafting
-      const {
-        allRecipes: craftingRecipes,
-        knownCount: knownCountCrafted,
-        craftedCount: craftedCount,
-      } = parseCrafting(jsonObj);
-
-      // Family & social
-      const { houseUpgradeLevel, spouse, children } = parseFamily(jsonObj);
-      const { fiveHeartCount, tenHeartCount, relationships } =
-        parseSocial(jsonObj);
-
-      // Monsters
-      const { deepestMineLevel, deepestSkullCavernLevel, monstersKilled } =
-        parseMonsters(jsonObj);
-
-      console.log("Parsed information!");
-
-      console.log("Uploading values to DB");
-      const dbstart = performance.now();
-      let response = await fetch("/api/kv", {
-        method: "PATCH",
-        body: JSON.stringify({
-          general: {
-            name,
-            timePlayed,
-            farmInfo,
-            moneyEarned,
-            questsCompleted,
-            uploadedFile: true,
-          },
-          stardrops: {
-            count: stardropsCount,
-            CF_Fair: stardrops.CF_Fair,
-            CF_Fish: stardrops.CF_Fish,
-            CF_Mines: stardrops.CF_Mines,
-            CF_Sewer: stardrops.CF_Sewer,
-            CF_Spouse: stardrops.CF_Spouse,
-            CF_Statue: stardrops.CF_Statue,
-            museumComplete: stardrops.museumComplete,
-          },
-          levels: {
-            player: levels["Player"],
-            farming: levels["Farming"],
-            fishing: levels["Fishing"],
-            foraging: levels["Foraging"],
-            mining: levels["Mining"],
-            combat: levels["Combat"],
-            maxLevelCount,
-          },
-          fish: {
-            totalCaught,
-            uniqueCaught,
-            ...allFish,
-          },
-          cooking: {
-            cookedCount,
-            knownCount,
-            ...allRecipes,
-          },
-          crafting: {
-            craftedCount,
-            knownCount: knownCountCrafted,
-            ...craftingRecipes,
-          },
-          mining: {
-            deepestMineLevel,
-            deepestSkullCavernLevel, // TODO: map through monstersKilled and add entry into DB for each
-            ...monstersKilled,
-          },
-          family: {
-            houseUpgradeLevel,
-            spouse: spouse ? spouse : "No spouse",
-            childrenLength: children ? children.length : 0,
-          },
-          social: {
-            fiveHeartCount,
-            tenHeartCount, // TODO: map through relationships and add entry into DB for each
-          },
-        }),
-      });
-      console.log(
-        "Completed DB uploads in",
-        (performance.now() - dbstart).toFixed(3),
-        "ms"
-      );
-
-      const elapsed = performance.now() - start;
-      console.log("Elapsed", elapsed.toFixed(3), "ms");
-      setCompletedTime((elapsed / 1000).toFixed(2));
-      setShowNotification(true);
     };
 
     reader.readAsText(file!);
